Tests chef

+++ code/model/ingredients/IngredientInventaire.java
@@ -26,7 +26,7 @@ public class IngredientInventaire {
      * @param unit  Unitée de mesure de l'ingrédient
      * @throws IngredientException Lance une exception si la quantité n'est pas valide
      */
-    public IngredientInventaire(Ingredient ingredient, int quantite, TypeUnit unit) throws IngredientException{
+    public IngredientInventaire(Ingredient ingredient, double quantite, TypeUnit unit) throws IngredientException{
         setIngredient(ingredient);
         setQuantite(quantite);
         setUnit(unit);
